Chore: move class to its own module
